tstest/integration: run Windows service test via the normal windows CI run
Updates #20381
